refactor: extract the Codex integration into @open-swe/codex
Everything specific to authenticating against Codex with a ChatGPT
subscription — the loopback credential broker, the request transport, and
the payload and header adaptations the backend requires — now lives in its
own package. `packages/agent` keeps only model selection and the agent
graph, and depends on it.
