+++ packages/agent/src/coding-agent.ts
@@ -10,17 +10,12 @@ import {
   resolveLocalProject,
 } from "@open-swe/workspace"
 
-import {
-  RESPONSES_LITE_HEADER,
-  adaptCodexPayload,
-  requiresResponsesLite,
-} from "./codex-payload.js"
+import { CHATGPT_API_BASE_URL, createOpenAiOAuthFetch } from "@open-swe/codex"
+
 import { createLocalSkillsMiddleware } from "./skills.js"
 
 const DEFAULT_MODEL_ID = "openai:gpt-5.6-sol"
 const OPENAI_EFFORTS = ["none", "low", "medium", "high", "xhigh"] as const
-const CHATGPT_API_BASE_URL = "https://chatgpt.com/backend-api/codex"
-const CODEX_ORIGINATOR = process.env.OPEN_SWE_CODEX_ORIGINATOR || "langchain"
 type OpenAIEffort = (typeof OPENAI_EFFORTS)[number]
 
 function isOpenAIEffort(value: unknown): value is OpenAIEffort {
@@ -28,87 +23,6 @@ function isOpenAIEffort(value: unknown): value is OpenAIEffort {
     typeof value === "string" &&
     (OPENAI_EFFORTS as readonly string[]).includes(value)
   )
-}
-
-interface OAuthBrokerCredentials {
-  access_token: string
-  account_id: string
-}
-
-function oauthBrokerConfig(env: NodeJS.ProcessEnv): {
-  url: URL
-  token: string
-} | null {
-  const rawUrl = env.OPEN_SWE_OPENAI_OAUTH_BROKER_URL
-  const token = env.OPEN_SWE_OPENAI_OAUTH_BROKER_TOKEN
-  if (!rawUrl || !token) return null
-  const url = new URL(rawUrl)
-  if (
-    url.protocol !== "http:" ||
-    url.hostname !== "127.0.0.1" ||
-    url.pathname !== "/token" ||
-    url.username ||
-    url.password
-  ) {
-    throw new Error("The local OpenAI credential broker address is invalid")
-  }
-  return { url, token }
-}
-
-export function createOpenAiOAuthFetch(
-  env: NodeJS.ProcessEnv,
-  fetchImpl: typeof fetch = globalThis.fetch
-): typeof fetch | null {
-  const broker = oauthBrokerConfig(env)
-  if (!broker) return null
-  return async (input, init) => {
-    const credentialResponse = await fetchImpl(broker.url, {
-      headers: { authorization: `Bearer ${broker.token}` },
-      signal: init?.signal,
-    })
-    if (!credentialResponse.ok) {
-      throw new Error("The local OpenAI session is unavailable; sign in again")
-    }
-    const credentials = (await credentialResponse.json()) as Partial<OAuthBrokerCredentials>
-    if (!credentials.access_token || !credentials.account_id) {
-      throw new Error("The local OpenAI session returned incomplete credentials")
-    }
-    const request = new Request(input, init)
-    const headers = new Headers(request.headers)
-    headers.set("authorization", `Bearer ${credentials.access_token}`)
-    headers.set("chatgpt-account-id", credentials.account_id)
-    // Codex serves its model allowlist per originator; an unrecognized value
-    // gets a stale list that no longer carries current models.
-    headers.set("originator", CODEX_ORIGINATOR)
-
-    const payload = await readJsonBody(request)
-    if (!payload) return fetchImpl(new Request(request, { headers }))
-    if (requiresResponsesLite(payload.model)) {
-      headers.set(RESPONSES_LITE_HEADER, "true")
-    }
-    return fetchImpl(
-      new Request(request.url, {
-        method: request.method,
-        headers,
-        body: JSON.stringify(adaptCodexPayload(payload)),
-        signal: request.signal,
-      })
-    )
-  }
-}
-
-async function readJsonBody(
-  request: Request
-): Promise<Record<string, unknown> | null> {
-  if (!["POST", "PATCH", "PUT"].includes(request.method)) return null
-  try {
-    const parsed: unknown = await request.clone().json()
-    return parsed && typeof parsed === "object" && !Array.isArray(parsed)
-      ? (parsed as Record<string, unknown>)
-      : null
-  } catch {
-    return null
-  }
 }
 
 export interface CodingAgentConfig {
