+++ packages/codex/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@open-swe/agent",
+  "name": "@open-swe/codex",
   "version": "0.0.0",
   "private": true,
   "type": "module",
@@ -20,13 +20,6 @@
     "build": "tsx ../../scripts/build-typescript-package.ts",
     "test": "vitest run",
     "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@langchain/core": "^1.2.9",
-    "@langchain/openai": "^1.5.1",
-    "@open-swe/codex": "workspace:*",
-    "@open-swe/workspace": "workspace:*",
-    "deepagents": "^1.13.1"
   },
   "devDependencies": {
     "@types/node": "^24.10.0",
